Reduce depth of git pull when checking out ports
An initial 'git clone' of the ports tree takes a long time. The checkout we
had before this change contained all the commit history, which would
normally not be necessary nor be interesting for people who simply want
to compile ports without doing any kind of development.

To reduce the clone time add the '--depth 1' parameter in this change.
Additionally, for people that do work on ports, a sentence about
removing that option was added.