Do not recommend git as an alternative to git
The link description would tell people to use git instead of git for
updating ports. Remove that cross reference as suggested in the PR. Note
that freebsd-update is eventually going away as well, but keep it here
for now and remove it in a separate change in the future.

Refrain from mentioning other third party tools in this section for keeping
consistency with the rest of the handbook.

PR:		    273359
Based on patch by:  grahamperrin@gmail.com