platforms/i386: report Pentium Pro as minimum supported x86 processor
As of FreeBSD 13.0[1] we default to i686 for 32-bit x86, and the
compiler will make use of the CMOV instruction.

It is still possible to compile FreeBSD/i386 for i486 or i586, but the
release artifacts are i686 and we don't regularly build or test i486 or
i586, and it may well be broken.

[1] https://www.freebsd.org/releases/13.0R/relnotes/

Reviewed by:	imp
Sponsored by:	The FreeBSD Foundation
Differential Revision: https://reviews.freebsd.org/D45734