website: Rename "More Events", "More Media" and "More News" to just "More"
The naming of these buttons is inconsistent - two of them just say
"More" and three say "More X". Because the section tittles are very
close and mention "X", I believe just "More" is better. This also allows
the removal of some i18n entries, including one that was inconsistent
(missing 1 of 4 entries).

Reviewed by:		0mp
Differential Revision:	https://reviews.freebsd.org/D45809