website: Relocate old status reports to the static folder
The files in question are pure HTML and were not converted to
Hugo/Asciidoctor formats.  In previous versions of Hugo, these were
merely treated as assets and were copied to the static folder during the
build process.

However, recent versions of Hugo are treating them as page resources,
similar to our .adoc files.

Moving these reports to the static folder will not only resolve this
issue but also allow us to upgrade our build servers to the latest
version of Hugo.

PR:		279815