documentation/themes: Enhance the book-menu feature
To activate the book menu, it is essential for the document to have a
showBookMenu parameter set as a boolean. However, not all documents use
a boolean; many utilize a string 'true' instead. This inconsistency
often arises from translations, where our current tool, po4a, does not
handle frontmatter correctly. This will ensure that the book menu
displays correctly regardless of whether showBookMenu is set as a
boolean or a string.