Release Notes: Hardware Compat: Move u3g to "Miscellaneous Networks"
At the FreeBSD vendor summit, it was revealed that modern modems
are a supported use case on FreeBSD via the extremely reliable
but little known u3g driver.

The Miscellaneous Networks section of the Hardware Release Notes
is usually empty.

Attempt to resolve both of cases -- the empty section and to make
people aware of u3g(4) -- by moving the cellular modem driver from
Ethernet to "Miscellaneous Networks" in the Release Hardware Notes.

Reviewed by:	bz
Differential Revision: https://reviews.freebsd.org/D47510