14.2/errata: ZFS home directory encryption
The adduser command is able to create encrypted home directories, but
we don't have any documentation advising users of how to decrypt their
home directories automatically when they log in.  Since this is a bit
of a foot-shooting issue, include it in the errata and hope that users
will read that page.

Sponsored by:	Amazon