14.2/relnotes: Tweak wording about BIOS support
We added back GZIP and BZIP support which were removed for 14.1 to try
to address size issues. However, lots of downstreams and minimalist
users use this .gz feature (and a few the .bz2 feature) so those were
added back. To cope with the size issues the BIOS loader has started
running into, especially with PXEBOOT, we instead moved to text-only
instead of the graphical boot loader.

Sponsored by:		Netflix