modify shasum check command
by default, sha256sum and sha512sum will look for any file mentioned in
the CHECKSUM file, which is all of them:

$ wc -l < CHECKSUM.SHA512-FreeBSD-14.1-RELEASE-amd64
10

with the current command, this will lead to 9 failures and 1 OK. the
fact that you specify the filename to be checked does not seem to mean
that is the *only* file that is checked.

for this reason, I suggest that the filename to be checked is removed,
and replaced with the flag --ignore-missing which is present in both
commands on both BSD and Linux.

this flag means that all files mentioned will be checked, just like
before, but output will only be returned for successful checks.

this assumes that the img file is in the same directory, but this was
already assumed in the previous version.