committers-guide: Drop notion of reference architectures
i386 can no longer be a reference architecture in 15.0 and later.
Instead, require new portable features to work on at least one Tier 1
architecture and drop the notion of a separate set of reference
architectures.

While here, FreeBSD has not been i386-only for so long that the
leading sentence now is outdated, so drop it.

Reviewed by:	imp, lwhsu
Differential Revision:	https://reviews.freebsd.org/D50620