handbook: ddd was removed in 2019
Remove obsolete ddd info: that port stopped working with gdb protocol
updates in about 2018, and was removed in 2019. It was abandoned
upstream around 2009 (last release). A new maintainer appeared in 2013,
but apparently didn't work out. Two new maintainers appeared in 2021 and
started new releases in 2023. A new port hasn't appeared, though, so
remove dangling references.

This whole section needs a rewrite, though, since dcons is so limited
(first 4GB only) and firewire is getting rare. All that should likely
just be removed. We need a good writeup on doing USB-C debugging with
the so-called USB-A to USB-A cable.

Sponsored by:		Netflix