handbook/network: Clarify use of Amnesiac
When FreeBSD calls itself Amnesiac, that is not actually the hostname.
It does that when there is no hostname set. It is not a substitute for
assigning a hostname. Setting a hostname or complaining when none is set
would introduce a regression for cases where the hostname is assigned by
the DHCP server. Attempt to clarify the handbook instead of changing the
behavior.

PR:		286847
Reported by:	jrtc27
Discussed with:	linimon