15.0: Developers have been reminded
An email went to the FreeBSD developers on July 11th reminding them
about the upcoming release.  It's time to start thinking about the
timelines for wrapping up works in progress to make sure they land
in time for 15.0; in particular, releng/15.0 is expected to branch
on October 3rd and any new features which haven't landed by then are
unlikely to be accepted into the release.