handbook: The freebsd-boot 512KB limit is for legacy BIOS
A question[1] on Reddit asked about the 512KB freebsd-boot limit due to
"current boot code limitations" in the context of EFI booting.  Be clear
that this is aobut legacy BIOS booting.

This section should have a larger rework for EFI, but this is at least
an improvement for the existing text.

[1] https://www.reddit.com/r/freebsd/comments/1mia9z6/boot_partition_over_512kb_works_fine/

Reviewed by:	imp
Sponsored by:	The FreeBSD Foundation
Differential Revision: https://reviews.freebsd.org/D51702