Avoid the use of "you" in the mail chapter
Rewrite sentences that contain 'you', which should be avoided according
to the FDP Primer. The resulting sentences are easier to understand and
often shorter. I did not change programlistings and other output as
these are from programs, which need to be changed first (out of scope of
this change).

Event:			    Oslo Hackathon 2025
Reviewed by:		    carlavilla
Differential Revision:	    https://reviews.freebsd.org/D52211