HW Relnotes template: Rework platforms for 15.0
Increase utility of the hardware release note by using the official
trade name of the architecture for it's subsection heading, followed
by the hyperlinked name of the freebsd project for that architecture.
In this hardware release note, I am also claiming a minimum RAM req.

This rewrite removes platforms which are no longer shipping in 15.0,
reorganizes the platforms subsections alpheabetically by project,
rewrites the introductory paragraps for each section to accomidate
for this (and past, really) reflow(s), and refreshes the powerpc
subsection with information from ppc@ and the platforms page.

I wish people using niche architectures to clean up their platforms
pages over the next two months. There is good information in there
drowning in a sea of stale info. This looks bad upon the project and
is subject to getting lost in a website rewrite.

Discussed with:	antranigv, cperciva, dch, emaste, michaeldexter