HW Relnotes template: Add sdhci and nda
Add the cherished NVMe storage and SD host controller interface
to the Hardware Release Note.

Those uninitiated in the hardware release note may wonder the
utility of listing a host controller interface in this document.

My mantra in the FreeBSD documentation project is:
"Your use case is not the center of the usecasiverse".

Earlier this week, I had the privilege of hearing the testimony of a
contractor who designed a sensor network for speedboats using FreeBSD.