Status/2025Q3/mac_do.adoc: Improve report
Add a preamble saying what mac_do(4)/mdo(1) is about and linking to the
manual pages and previous status reports.

Improve wording, simplify.

Clarify that only the work on mdo(1) to enable fine-grained credentials
transition requests has been committed, and mention in which releases it
will appear.  Mention that the other points will eventually be
committed, but may not appear in 15.0

Add myself as a point of contact, and mention sponsoring by the FreeBSD
Foundation.

Sponsored by:   The FreeBSD Foundation