Status/2025Q3/group-changes.adoc: Add report
This is a report concerning credentials' group-related changes and in
particular the project of improving the behavior of
setgroups(2)/getgroups(2) and initgroups(3) to avoid security pitfalls
and be compatible with most other open-source systems.

It follows the similarly named report for T2 2025, which is linked from
this new one.

Sponsored by:   The FreeBSD Foundation