Contributing article: Indicate it is for Bugzilla
This Article was written long before Phabricator and Pull Requests were
accepted ways to submit change requests or bug issues to FreeBSD.

As a first step to bringing it up to date, disambiguate the "PR" term,
and spell it out as "Bugzilla Problem Request".

There are many more problems here to be solved.

Reviewed by:		imp, pauamma_gundo.com, ziaee
Differential Revision:	https://reviews.freebsd.org/D53638