Pr-Guidelines: Indicate this is for Bugzilla
This article has not been updated in years.  There is so much wrong,
that it will need to be tackled in multiple patches.

This first patch simply states that the article is Bugzilla-specific,
and that any use of "PR" means "Bugzilla Problem Report".  In a few
cases, this is now spelled out explicitly.

Discused with:		imp, lwhsu, ziaee
Differential Revision:	https://reviews.freebsd.org/D53645