releases/15.0R/relnotes: Fix and expand paragraphs on SMBIOS
Improve description of what was changed.

Move the change in smbios(3) under the kernel driver section, this one
is not related to the boot loaders.

Stop referring to libsa(3) for changes in the boot loaders, which most
probably does not speak to the intended public.  Instead, clearly say
whether the BIOS or EFI loader is concerned.

Re-order changes chronologically, except for one of them in order to
ease understanding of the descriptions.

Fix commit hashes (they pointed to MFC commits to stable/14).

While here, in the changed paragraphs, fix punctuation around the commit
links and "Sponsored by The FreeBSD Foundation" lines to match what
I think was common practice in earlier release notes.

Sponsored by:   The FreeBSD Foundation