releases/15.0R/relnotes: Expand coverage of mac_do(4)/mdo(1)
All important messages should be conveyed now.

Use a less telegraphic style for first sentences.

Fix pre-existing commit hashes (they pointed to MFC commits to stable/14).

While here, in the changed paragraphs, fix punctuation around the commit
links and "Sponsored by The FreeBSD Foundation" lines, though, comparing
with release notes for 14.0 and 13.0, this may not be the final style we
want, and anyway the whole file will have to be revised for uniformity.