handbook/desktop: Refresh Plasma
I tried running plasma wayland on 15.0, it works great! CPU usage
is a tad high, but I'm impressed! Freshen the branding and language
in this chapter to highlight this. Remove a stale note about adjusting
message sizes to what is now the default. Tell the reader how to search
the kde-applications category. Add instructions to launch with wayland.

Reviewed by:		carlavilla (previous), makc
Differential Revision:	https://reviews.freebsd.org/D54207