15.0/relnote: Organize Upgrading section
Judging from the mailing list discussion, this section needs organized
better. While here, sneak in a bit of wordsmithing. We still have a lot
of work to do improving the instructions for upgrading systems installed
with packages, but we have to keep moving.

Discussed with:		cperciva (it's still ok to organize better)
Discussed with:		ivy (s/distribution/& sets/)
Discussed with:		jrm (upgrading from source wordsmithing)
Reviewed by:		dch (previous, fixed a typo since)
Differential Revision:	https://reviews.freebsd.org/D54181